PLATUI-3143: Combined BrowserDriver methods into BasePage, removed duplicate code
